final commit

+++ src/main.cpp
@@ -16,9 +16,9 @@ using namespace hctg;
 C_Application::T_PressedKey g_PressedKeys = 0;
 
 
-const int k_ScreenWidth = hctd::Resource::inst().getValue(hctd::SCREEN_WIDTH);
-const int k_ScreenHeight = hctd::Resource::inst().getValue(hctd::SCREEN_HEIGHT);
-const int k_FrameRate = hctd::Resource::inst().getValue(hctd::FRAME_RATE);
+const int k_ScreenWidth = (int)hctd::Resource::inst().getValue(hctd::SCREEN_WIDTH);
+const int k_ScreenHeight = (int)hctd::Resource::inst().getValue(hctd::SCREEN_HEIGHT);
+const int k_FrameRate = (int)hctd::Resource::inst().getValue(hctd::FRAME_RATE);
 
 
 //----------------------------
